# src/backend/openapi.yaml
openapi: 3.0.0
info:
  title: Nightfall API
  description: API for Nightfall theme park experiences, reservations, and user management
  version: 1.0.0
servers:
  - url: http://localhost:5080/api
    description: Local development server
  - url: https://api.nightfall.example.com
    description: Production server
paths:
...
  /reservations:
    get:
      summary: Get user's reservations
      description: Retrieve all reservations for the authenticated user
      tags:
        - Member
      security:
        - bearerAuth: []
      responses:
        '200':
          description: A list of reservations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Reservation'
        '401':
          description: Unauthorized
        '404': 
          description: No reservations found
        '500':
          description: Internal server error

    post:
      summary: Create a new reservation
      description: Reserve an experience for the authenticated user
      tags:
        - Member
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReservationInput'
      responses:
        '201':
          description: Reservation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reservation'
        '400':
          description: Invalid input
        '401':
          description: Unauthorized
        '404':
          description: Experience not found
        '409':
          description: Reservation conflict (e.g., capacity exceeded)
        '500':
          description: Internal server error
  /reservations/{id}:
    delete:
      summary: Cancel a reservation
      description: Cancel a reservation for the authenticated user
      tags:
        - Member
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          description: ID of the reservation to cancel
          schema:
            type: integer
      responses:
        '204':
          description: Reservation cancelled successfully
        '401':
          description: Unauthorized
        '403':
          description: Forbidden (e.g., not the reservation owner)
        '404':
          description: Reservation not found
        '409':
          description: Cancellation not allowed (e.g., too late)
        '500':
          description: Internal server error
components:
  schemas:
...
    Reservation:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the reservation
        experience_id:
          type: integer
          description: ID of the reserved experience
        user_id:
          type: integer
          description: ID of the user who made the reservation
        date:
          type: string
          format: date-time
          description: Date and time of the reservation
        participants:
          type: integer
          description: Number of participants

      required:
        - id
        - experience_id
        - user_id
        - date
        - participants
    ReservationInput:
      type: object
      properties:
        experience_id:
          type: integer
          description: ID of the experience to reserve
        date:
          type: string
          format: date-time
          description: Date and time of the reservation
        participants:
          type: integer
          description: Number of participants
      required:
        - experience_id
        - date
        - participants
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT